refactor: app 모듈에서 Injector 이용
