+++ data/src/main/java/com/github/dodobest/data/GithubRepositoryImpl.kt
@@ -1,11 +1,11 @@
 package com.github.dodobest.data
 
-import com.github.dodobest.data.retrofit.RetrofitManager
 import com.github.dodobest.domain.GithubData
 import com.github.dodobest.domain.GithubRepository
 import com.github.dodobest.domain.RESPONSE_STATE
+import com.github.dodobest.domain.RetrofitManager
 
-class GithubRepositoryImpl(
+internal class GithubRepositoryImpl(
     private val retrofitManager: RetrofitManager
 ) : GithubRepository {
     override fun getRepositories(completion: (RESPONSE_STATE, List<GithubData>) -> Unit) {
