+++ data/src/main/java/com/github/dodobest/data/utils/Injector.kt
@@ -1,11 +1,22 @@
 package com.github.dodobest.data.utils
 
 import com.github.dodobest.data.GithubRepositoryImpl
-import com.github.dodobest.data.retrofit.RetrofitManager
+import com.github.dodobest.data.retrofit.RetrofitClient
+import com.github.dodobest.data.retrofit.RetrofitManagerImpl
 import com.github.dodobest.domain.GithubRepository
+import com.github.dodobest.domain.RetrofitManager
+import retrofit2.Retrofit
 
 object Injector {
-    fun provideTaskRepository(): GithubRepository {
-        return GithubRepositoryImpl(RetrofitManager())
+    fun provideGithubRepository(retrofitManager: RetrofitManager): GithubRepository {
+        return GithubRepositoryImpl(retrofitManager)
+    }
+
+    fun provideRetrofitManager(retrofit: Retrofit): RetrofitManager {
+        return RetrofitManagerImpl(retrofit)
+    }
+
+    fun provideRetrofit(): Retrofit {
+        return RetrofitClient.getRetrofit()
     }
 }
